playwright-test-demo/requirements.txt
pytest==8.3.4
playwright==1.49.1
pytest-playwright==0.6.2
python-dotenv==1.0.1
allure-pytest==2.13.5
